# fix: reduce allocations in chain traversal