## Summary of changes

<!-- Please write a comprehensive summary of your changes and what was the motivation behind them -->

This PR tries to reduce heap allocations in chain traversal (`fn stream_chain`) by replacing `Vec` with `SmallVec`. This would benefit chain export and snapshot validatation, etc.

BTW, I believe it could be further improved by refactoring `fn extract_cids` into returning an iterator of CIDs

Perf gains on my laptop:
calibnet:  ~4m40s -> ~3m50s
mainnet: ~47m -> ~37m

```
PR branch:
➜  snapshots forest-tool -V && forest-tool benchmark graph-traversal forest_snapshot_calibnet_2023-07-30_height_780000.forest.car.zst
forest-filecoin 0.29.0+git.0c00fb4c
  traversed 2.65 GiB at 65.20 MiB/s in 00:00:41                                                                         ➜  snapshots forest-tool -V && forest-tool benchmark graph-traversal forest_snapshot_calibnet_2023-07-30_height_780000.forest.car.zst
forest-filecoin 0.29.0+git.0c00fb4c
  traversed 2.65 GiB at 66.89 MiB/s in 00:00:40                                                                         ➜  snapshots forest-tool -V && forest-tool benchmark graph-traversal forest_snapshot_calibnet_2023-07-30_height_780000.forest.car.zst
forest-filecoin 0.29.0+git.0c00fb4c
  traversed 2.65 GiB at 66.89 MiB/s in 00:00:40               
  
main branch:
➜  snapshots forest-tool -V && forest-tool benchmark graph-traversal forest_snapshot_calibnet_2023-07-30_height_780000.forest.car.zst
forest-filecoin 0.29.0+git.15cdbb31
  traversed 2.65 GiB at 64.23 MiB/s in 00:00:42                                                                         ➜  snapshots forest-tool -V && forest-tool benchmark graph-traversal forest_snapshot_calibnet_2023-07-30_height_780000.forest.car.zst
forest-filecoin 0.29.0+git.15cdbb31
  traversed 2.65 GiB at 66.59 MiB/s in 00:00:40                                                                         ➜  snapshots forest-tool -V && forest-tool benchmark graph-traversal forest_snapshot_calibnet_2023-07-30_height_780000.forest.car.zst
forest-filecoin 0.29.0+git.15cdbb31
  traversed 2.65 GiB at 62.37 MiB/s in 00:00:43             
```

Changes introduced in this pull request:

-

## Reference issue to close (if applicable)

<!-- Include the issue reference this pull request is connected to -->
<!-- See more keywords here https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword -->
<!--(e.g. Closes #1)-->

Closes

## Other information and links

<!-- Add any other context about the pull request here. Those might be helpful links based on your investigation, relevant commits from this or other repositories or anything else -->

## Change checklist

<!-- Please add a changelog entry for your change if needed. -->
<!-- Follow this format https://keepachangelog.com/en/1.0.0/ -->

- [x] I have performed a self-review of my own code,
- [x] I have made corresponding changes to the documentation. All new code adheres to the team's [documentation standards](https://github.com/ChainSafe/forest/wiki/Documentation-practices),
- [x] I have added tests that prove my fix is effective or that my feature works (if possible),
- [x] I have made sure the [CHANGELOG][1] is up-to-date. All user-facing changes should be reflected in this document.

<!-- Thank you 🔥 -->

[1]: https://github.com/ChainSafe/forest/blob/main/CHANGELOG.md


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- Performance
  - Reduced memory allocations in CID extraction for faster, more efficient CBOR decoding and benchmarking workflows.

- API Changes
  - Updated the CID extraction function to return a small-vector type instead of a standard vector; update integrations to use the new return type.

- Refactor
  - Minor cleanup to leverage type inference in benchmarking logic with no behavioral changes.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->